requirements.txt
amqp==5.2.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
billiard==4.2.0
boto3==1.34.102
botocore==1.34.102
celery==5.4.0
certifi==2024.2.2
cffi==1.16.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.3
Django==5.0.6
django-celery-beat==2.6.0
django-filter==24.2
django-minio-storage==0.5.7
django-timezone-field==6.1.0
djangorestframework==3.15.1
drf-yasg==1.21.7
flower==2.0.1
humanize==4.9.0
inflection==0.5.1
jmespath==1.0.1
kombu==5.3.7
minio==7.2.7
packaging==24.0
pillow==10.3.0
prometheus_client==0.20.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
pycparser==2.22
pycryptodome==3.20.0
python-crontab==3.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
redis==5.0.4
s3transfer==0.10.1
setuptools==69.5.1
six==1.16.0
sqlparse==0.5.0
tornado==6.4
typing_extensions==4.11.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
gunicorn==20.1.0